Refactor code into namespaces
